Adding CXX allocation problem

+++ Pragma_Examples/OpenMP/Fortran/optimization/Allocations/3_memorypool/memorypool.f90
@@ -22,10 +22,10 @@ program main
     integer :: n = 10000000
     integer,parameter :: Niter = 10
     ! Input vectors and Output vector
-    real64,dimension(:),pointer,contiguous :: a, b, c
+    real(kind=real64),dimension(:),pointer,contiguous :: a, b, c
     integer :: i,iter
-    real64 :: sum
-    real64 :: startt, endt
+    real(kind=real64) :: sum
+    real(kind=real64) :: startt, endt
 
     startt=omp_get_wtime()
